Remedy clang compiler warnings, format all
Summary: - Fixes for warnings introduced in Batch Normalization commit

Test Plan: 37/37 tests pass (OS/X)

Reviewers: simonl

Reviewed By: simonl

Differential Revision: https://phabricator.sourcevertex.net/D8133

+++ willow/include/poponnx/op/batchnorm.hpp
@@ -29,7 +29,7 @@ public:
   // Attributes
   float getEpsilon() { return epsilon; }
   float getMomentum() { return momentum; }
-  int getSpatial() { return spatial; }
+  int64_t getSpatial() { return spatial; }
 
   bool isTraining() { return training; }
 
@@ -37,7 +37,7 @@ private:
   bool training = false;
   float epsilon;
   float momentum;
-  int spatial;
+  int64_t spatial;
 };
 
 class BatchNormGradOp : public Op {
